remove tiled files
