--- tiled/pokemon_tileset_ 4.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.4" tiledversion="1.4.1" name="pokemon_tileset_ 4" tilewidth="32" tileheight="32" tilecount="5208" columns="8">
- <image source="src/assets/tilesets/pokemon_tileset_from_public_tiles_32x32_by_chaoticcherrycake_dab2byf (3).png" width="256" height="20832"/>
-</tileset>
