--- tiled/pokemon_tileset_cropped.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.4" tiledversion="1.4.1" name="pokemon_tileset_cropped" tilewidth="32" tileheight="32" tilecount="4096" columns="8">
- <image source="src/assets/tilesets/pokemon_tileset_cropped.png" width="256" height="16384"/>
-</tileset>
