--- tiled/pokemon_tileset_cropped3.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.4" tiledversion="1.4.1" name="pokemon_tileset_cropped3" tilewidth="21" tileheight="21" margin="16" tilecount="8569" columns="11">
- <image source="src/assets/tilesets/pokemon_tileset_cropped.png" width="256" height="16384"/>
-</tileset>
