--- tiled/test.tmx
@@ -1,150 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<map version="1.4" tiledversion="1.4.1" orientation="orthogonal" renderorder="right-down" width="25" height="25" tilewidth="32" tileheight="32" infinite="0" nextlayerid="3" nextobjectid="1">
- <tileset firstgid="1" name="pokemon_tileset_ 4" tilewidth="32" tileheight="32" tilecount="5208" columns="8">
-  <image source="src/assets/tilesets/pokemon_tileset_from_public_tiles_32x32_by_chaoticcherrycake_dab2byf (3).png" width="256" height="20832"/>
-  <tile id="320">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="10.4348" y="31.8535">
-     <polygon points="0,0 -2.56293,-2.1968 5.12586,-11.0755 13.4554,-2.746 11.8078,0"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="321">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="9.24485" y="31.762">
-     <polygon points="0,0 -1.00686,-1.55606 6.59039,-10.4348 14.5538,-1.6476 13.4554,0"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="328">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="16.0183" y="18.2151"/>
-    <object id="2" x="15.9268" y="10.2517"/>
-   </objectgroup>
-  </tile>
-  <tile id="329">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="16.0183" y="18.4897"/>
-    <object id="2" x="16.476" y="9.88558"/>
-   </objectgroup>
-  </tile>
-  <tile id="3096">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="11.5332" y="31.3043">
-     <polygon points="0,0 19.9542,-13.1808 20.2288,0.457666"/>
-    </object>
-    <object id="2" x="18.6728" y="27.2769">
-     <polygon points="0,0 3.02059,-2.1968 -4.02746,-1.373"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="3097">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="0.549199" y="18.4897">
-     <polygon points="0,0 11.0755,-6.68192 16.0183,-6.59039 26.6362,-1.28146 27.5515,12.8146 -0.183066,13.1808"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="3104">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="21.4188" y="9.97712"/>
-    <object id="2" x="21.6934" y="21.968"/>
-   </objectgroup>
-  </tile>
-  <tile id="3105">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="15.286" y="15.6522"/>
-    <object id="2" x="10.2517" y="10.1602"/>
-    <object id="3" x="9.79405" y="21.5103"/>
-    <object id="4" x="21.5103" y="21.7849" width="0.183066" height="0.0915332"/>
-    <object id="5" x="21.968" y="10.3432"/>
-   </objectgroup>
-  </tile>
-  <tile id="3106">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="17.5744" y="0.457666">
-     <polygon points="0,0 -0.823799,26.087 4.39359,31.3959 -17.6659,31.2128 -17.1167,-0.366133"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="3112">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="15.286" y="31.2128">
-     <polygon points="0,0 -4.39359,-30.8467 16.5675,-31.0297 16.659,-3.29519 0.640732,0.0915332"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="3113">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="0.915332" y="28.0092">
-     <polygon points="0,0 30.2975,0.640732 30.8467,-27.5515 -0.366133,-27.5515"/>
-    </object>
-   </objectgroup>
-  </tile>
-  <tile id="3114">
-   <objectgroup draworder="index" id="2">
-    <object id="1" x="17.5744" y="0.0915332">
-     <polygon points="0,0 -0.2746,9.97712 -4.11899,13.5469 -4.39359,30.5721 -17.3913,30.8467 -17.2082,0.366133"/>
-    </object>
-   </objectgroup>
-  </tile>
- </tileset>
- <layer id="1" name="BottomLayer" width="25" height="25">
-  <data encoding="csv">
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,
-707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707,707
-</data>
- </layer>
- <layer id="2" name="TopLayer" width="25" height="25">
-  <data encoding="csv">
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,3097,3098,3099,3097,3098,3099,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,3105,3106,3107,3105,3106,3107,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,3113,3114,3115,3113,3114,3115,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,321,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,329,0,0,0,0,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,321,0,0,0,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,321,0,329,0,0,321,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,321,0,0,329,0,321,0,0,329,0,321,0,
-0,0,0,0,0,0,0,0,321,0,0,0,0,329,0,0,0,321,329,0,321,0,0,329,0,
-0,0,0,0,0,0,0,0,329,0,0,0,0,0,0,0,0,329,0,321,329,321,321,0,0,
-0,0,0,0,0,0,0,0,0,321,0,0,0,321,0,0,1037,1037,0,329,0,329,329,0,0,
-0,0,0,0,0,0,0,0,0,329,0,0,0,329,0,1037,1037,1037,1037,0,321,321,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1037,1037,1037,1037,321,329,329,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,321,0,1037,1037,0,0,329,321,0,0,321,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,329,0,0,321,0,0,321,329,321,0,329,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,329,0,0,329,0,329,0,0,0,
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,329,0,0,0,0,0,0,0,0
-</data>
- </layer>
-</map>
