--- tiled/tiled_project.tiled-project
@@ -1,9 +1,0 @@
-{
-    "automappingRulesFile": "",
-    "commands": [
-    ],
-    "extensionsPath": "extensions",
-    "folders": [
-    ],
-    "objectTypesFile": ""
-}
