--- tiled/tiled_project.tiled-session
@@ -1,54 +1,0 @@
-{
-    "Map/SizeTest": {
-        "height": 4300,
-        "width": 2
-    },
-    "activeFile": "test.tmx",
-    "expandedProjectPaths": [
-    ],
-    "fileStates": {
-        "": {
-            "scaleInDock": 1
-        },
-        "pokemon_tileset_2.tsx": {
-            "scaleInDock": 1,
-            "scaleInEditor": 1
-        },
-        "pokemon_tileset_cropped.tsx": {
-            "scaleInDock": 1,
-            "scaleInEditor": 1
-        },
-        "pokemon_tileset_cropped3.tsx": {
-            "scaleInDock": 1,
-            "scaleInEditor": 1
-        },
-        "test.tmx": {
-            "scale": 0.12051361386138613,
-            "selectedLayer": 0,
-            "viewCenter": {
-                "x": 1576.5853658536587,
-                "y": 1597.3299101412067
-            }
-        }
-    },
-    "map.lastUsedExportFilter": "JSON map files (*.json)",
-    "map.lastUsedFormat": "tmx",
-    "map.layerDataFormat": null,
-    "openFiles": [
-        "test.tmx",
-        "pokemon_tileset_cropped.tsx"
-    ],
-    "project": "tiled_project.tiled-project",
-    "recentFiles": [
-        "pokemon_tileset_cropped3.tsx",
-        "pokemon_tileset_cropped.tsx",
-        "pokemon_tileset_2.tsx",
-        "test.tmx"
-    ],
-    "tileset.lastUsedFormat": "tsx",
-    "tileset.margin": 0,
-    "tileset.tileSize": {
-        "height": 32,
-        "width": 32
-    }
-}
